+++ internal/proxy/util.go
@@ -1,4 +1,4 @@
-package util
+package proxy
 
 import (
 	"context"
